% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInteractions.R
\name{allInt}
\alias{allInt}
\title{allInt}
\usage{
allInt(mat, type = "lollipop", top = 0, ...)
}
\arguments{
\item{mat}{A matrix of values to be plotted. Either added by the user or created using the prepFunc() function.}

\item{type}{The type of plot to display, either "lollipop", "barplot", or "circleBar".}

\item{top}{A value set by the user to only display the top x amount variables.}

\item{...}{Not currently implemented}
}
\description{
Plots a Heatmap-tyle display showing Variable Importance and Variable Interaction.
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr random forest model:
library(mlr)
library(randomForest)
aqRgrTask  <- makeRegrTask(data = aq, target = "Ozone")
aqRegrLrn <- makeLearner("regr.randomForest")
aqMod <- train(aqRegrLrn, aqRgrTask)

# Create plot:
allInt(aqRgrTask, aqMod)

}
