% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInteractions.R
\name{allInteractionsPlot}
\alias{allInteractionsPlot}
\title{allInteractionsPlot}
\usage{
allInteractionsPlot(mat, plotType = NULL, top = NULL, ...)
}
\arguments{
\item{mat}{A matrix of values to be plotted. Either added by the user or created using the prepFunc() function.}

\item{top}{A value set by the user to only display the top x amount variables.}

\item{...}{Not currently implemented}

\item{type}{The type of plot to display, either "lollipop", "barplot", or "circleBar".}
}
\description{
Plots all 2-way interactions
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr ranger model:
library(mlr3)
library(mlr3learners)
library(ranger)
aq_Task = TaskRegr$new(id = "airQ", backend = aq, target = "Ozone")
aq_lrn = lrn("regr.ranger", importance = "permutation")
aq_Mod <- aq_lrn$train(aq_Task)

# Create matrix
myMat <- vividMatrix(task = aq_Task,  model = aq_Mod)

# Create plot:
plot(myMat, type = "allInteractions", plotType = "barplot")

}
