% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkPlot.R
\name{importanceNet}
\alias{importanceNet}
\title{Title importanceNet}
\usage{
importanceNet(
  task,
  model,
  method = "randomForest_importance",
  thresholdValue = 0,
  label = T,
  minInt = 0,
  maxInt = 1,
  minImp = 0,
  maxImp = impLegend,
  cluster = F,
  ...
)
}
\arguments{
\item{task}{Task created from the mlr package, either regression or classification.}

\item{model}{Any machine learning model.}

\item{method}{A list of variable importance methods to be set by the user. These can include any of the importance methods contained within the mlr package. The default is method = randomForest.}

\item{thresholdValue}{A value chosen by the user which will show all the edges with weights (i.e., the interacions) above that value. For example, if thresholdValue = 0.2, then only the the interacions greater than 0.2 will be displayed.}

\item{label}{If label = TRUE the numerical value for the interaction strength will be displayed.}

\item{minInt}{Minimum interaction strength to be displayed on the legend.}

\item{maxInt}{Maximum interaction strength to be displayed on the legend.}

\item{minImp}{Minimum importance value to be displayed on the legend.}

\item{maxImp}{Maximum importance value to be displayed on the legend.}

\item{cluster}{If cluster = TRUE, then the data is clustered in groups.}
}
\value{
A newtwork style plot displaying interaction strength between variables on the edges and variable importance on the nodes
}
\description{
Title importanceNet
}
