% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotInteraction.R
\name{interactionPlot}
\alias{interactionPlot}
\title{interactionPlot}
\usage{
interactionPlot(task, model, type = "lollipop")
}
\arguments{
\item{task}{Task created from the mlr package, either regression or classification.}

\item{model}{Any machine learning model.}

\item{type}{The type of plot to display, either "lollipop" (default), "barplot", or "circleBar"}
}
\description{
Plots a selected variable against all other variables in a model
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr ranger model:
library(mlr3)
library(mlr3learners)
library(ranger)
aq_Task = TaskRegr$new(id = "airQ", backend = aq, target = "Ozone")
aq_lrn = lrn("regr.ranger", importance = "permutation")
aq_Mod <- aq_lrn$train(aq_Task)

# Create plot:
interactionPlot(aq_Task, aq_Mod)

}
