% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdpPairs.R
\name{pdpPairs}
\alias{pdpPairs}
\title{pdpPairs}
\usage{
pdpPairs(task, model, cols = rev(sequential_hcl(20, "Blues3")), ...)
}
\arguments{
\item{task}{Task created from the mlr package, either regression or classification.}

\item{model}{Any machine learning model.}

\item{cols}{Colour of partial dependece plot.}

\item{...}{Not currently implemented}
}
\value{
A pairs style plot displaying the partial dependence.
}
\description{
Creates a plot of the partial dependence of each of the variables in pairs plot style matrix
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr random forest model:
library(mlr)
library(randomForest)
aqRgrTask  <- makeRegrTask(data = aq, target = "Ozone")
aqRegrLrn <- makeLearner("regr.randomForest")
aqMod <- train(aqRegrLrn, aqRgrTask)

# Create plot:
pdpPairs(aqRgrTask , aqMod)

}
