% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zPlot.R
\name{pdpZenplot}
\alias{pdpZenplot}
\title{Create a zenplot displaying partial dependence values.}
\usage{
pdpZenplot(
  task,
  model,
  zpath = NULL,
  method = "pdp",
  noCols = c("letter", "square", "A4", "golden", "legal"),
  zenMethod = c("tidy", "double.zigzag", "single.zigzag", "rectangular"),
  colLow = "#D7191C",
  colMid = "#FFFFBF",
  colHigh = "#2B83BA",
  fitlims = NULL,
  gridsize = 10,
  class = 1,
  ...
)
}
\arguments{
\item{task}{Task created from the mlr3 package, either regression or classification.}

\item{model}{A machine learning model created from mlr3 task and learner.}

\item{zpath}{A zenpath created from calcZpath. see \code{\link[zenplots]{zenpath}} from the
\code{\link[zenplots]{zenplots}} package for more details.}

\item{method}{"pdp" (default) or "ale"}

\item{noCols}{number of columns of 2d plots (>= 1) or one of "letter", "square", "A4", "golden" or "legal"
in which case a similar layout is constructed. See ?zenplot}

\item{zenMethod}{String indicating the layout of the zigzag plot. The available methods are:
"tidy": more tidied-up double.zigzag (slightly more compact placement of plots towards the end).
"double.zigzag": zigzag plot in the form of a flipped “S”. Along this path, the plots are placed in the form of an “S” which is rotated counterclockwise by 90 degrees.
"single.zigzag": zigzag plot in the form of a flipped “S”.
"rectangular": plots that fill the page from left to right and top to bottom. This is useful (and most compact) for plots that do not share an axis.}

\item{colLow}{Colour to be used for low values.}

\item{colMid}{Colour to be used for mid values.}

\item{colHigh}{Colour to be used for low values.}

\item{fitlims}{If supplied, should be a numeric vector of length 2, specifying the fit range.}

\item{gridsize}{for the pdp/ale plots, defaults to 10.}

\item{class}{For a classification model, show the probability of this class. Defaults to 1.}
}
\value{
A zenplot of partial dependence values.
}
\description{
Constructs a zigzag expanded navigation plot (zenplot) displaying partial dependence values.
}
\examples{
# Load in the data:
aq <- na.omit(airquality)*1.0

# Run an mlr3 ranger model:
library(mlr3)
library(mlr3learners)
library(ranger)
ozonet  <- TaskRegr$new(id = "airQ", backend = aq, target = "Ozone")
ozonel <- lrn("regr.ranger", importance = "permutation")
ozonef  <- ozonel$train(ozonet)

# Create matrix
viv <- vividMatrix(ozonet, ozonef)

# Calculate Zpath:
zpath<-calcZpath(viv,.8)
zpath

# Create graph:
pdpZenplot(ozonet, ozonef, zpath=zpath)

}
