% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{reorderMatrix}
\alias{reorderMatrix}
\title{reorderMatrix}
\usage{
reorderMatrix(matrix)
}
\arguments{
\item{matrix}{A matrix to be reordered.}
}
\value{
A matrix of reordered values
}
\description{
Used to reorder a square matrix so that values of high importance and
interaction strength are pushed to the top left of the matrix.
}
\examples{
library(vivid)

# Create data
myData <- genFriedman(10, 100)

# Run an mlr ranger model:
library(mlr3)
library(mlr3learners)
library(ranger)

Task = TaskRegr$new(id = "Fried", backend = myData, target = "y")
lrn = lrn("regr.ranger", importance = "permutation")
Mod <- lrn$train(Task)

# Create matrix
myMat <- vividMatrix(task = Task, model = Mod, gridsize = 100, reorder = FALSE)

myMat <- reorderMatrix(myMat)

}
