% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionMatrix.R
\name{vividMatrix}
\alias{vividMatrix}
\title{vividMatrix}
\usage{
vividMatrix(
  task,
  model,
  gridSize = 10,
  normalize = FALSE,
  n_max = 1000,
  seed = NULL,
  sqrt = FALSE,
  ...
)
}
\arguments{
\item{task}{Task created from the mlr package, either regression or classification.}

\item{model}{Any machine learning model.}

\item{gridSize}{The size of the grid for evaluating the predictions.}

\item{normalize}{Should the variances explained be normalized? Default is FALSE.}

\item{n_max}{Maximum number of data rows to consider.}

\item{seed}{An integer random seed used for subsampling.}

\item{sqrt}{In order to reproduce Friedman's H statistic, resulting values are root transformed. Set to FALSE if squared values should be returned.}
}
\value{
A matrix of values
}
\description{
Creates a matrix displaying Variable importance on the diagonal
 and Variable Interaction on the off-diagonal.
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr ranger model:
library(mlr3)
library(mlr3learners)
library(ranger)
aq_Task = TaskRegr$new(id = "airQ", backend = aq, target = "Ozone")
aq_lrn = lrn("regr.ranger", importance = "permutation")
aq_Mod <- aq_lrn$train(aq_Task)

# Create matrix
myMat <- vividMatrix(model = aq_Mod)

# Create graph:
plot(myMat, type = "heatMap")

}
