% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zPlot.R
\name{zPlot}
\alias{zPlot}
\title{zPlot}
\usage{
zPlot(
  task,
  model,
  method = "pdp",
  vars = NULL,
  colLow = "#D7191C",
  colMid = "#FFFFBF",
  colHigh = "#2B83BA",
  fitlims = NULL,
  gridsize = 10,
  class = 1
)
}
\arguments{
\item{task}{Task created from the mlr package, either regression or classification.}

\item{model}{Any machine learning model from mlr3.}

\item{method}{"pdp" (default) or "ale"}

\item{vars}{Variables to plot. Defaults to all predictors.}

\item{colLow}{Colour to be used for low values.}

\item{colMid}{Colour to be used for mid values.}

\item{colHigh}{Colour to be used for low values.}

\item{fitlims}{If supplied, should be a numeric vector of length 2, specifying the fit range.}

\item{gridsize}{for the pdp/ale plots, defaults to 10.}

\item{class}{For a classification model, show the probability of this class. Defaults to 1.}

\item{...}{Not currently implemented.}
}
\value{
A zenplot displaying the 2-way partial dependence.
}
\description{
@description Create a zenplot displaying the 2-way partial dependence of each of the variables.
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr ranger model:
library(mlr3)
library(mlr3learners)
library(ranger)
aq_Task = TaskRegr$new(id = "airQ", backend = aq, target = "Ozone")
aq_lrn = lrn("regr.ranger", importance = "permutation")
aq_Mod <- aq_lrn$train(aq_Task)

# Create plot:
zPlot(task = aq_Task, model = aq_Mod)

}
