% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInteractions.R
\name{allInt}
\alias{allInt}
\title{allInt}
\usage{
allInt(task, model, type = "lollipop", top = 0, ...)
}
\arguments{
\item{task}{Task created from the mlr package, either regression or classification.}

\item{model}{Any machine learning model.}

\item{type}{The type of plot to display, either "lollipop", "barplot", or "circleBar"}

\item{top}{A value set by the user to only display the top x amount variables}

\item{...}{Not currently implemented}
}
\description{
Plots a Heatmap-tyle display showingVariable Importance and Variable Interaction
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr random forest model:
library(mlr)
library(randomForest)
aqRgrTask  <- makeRegrTask(data = aq, target = "Ozone")
aqRegrLrn <- makeLearner("regr.randomForest")
aqMod <- train(aqRegrLrn, aqRgrTask)

# Create plot:
allInt(aqRgrTask, aqMod)

}
