% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotImportance.R
\name{importancePlot}
\alias{importancePlot}
\title{importancePlot}
\usage{
importancePlot(mat, type = "lollipop", ...)
}
\arguments{
\item{type}{The type of plot to display, either "lollipop" (default), "barplot", or "circleBar"}

\item{...}{Not currently implemented}

\item{task}{Task created from the mlr package, either regression or classification.}

\item{model}{Any machine learning model.}

\item{method}{A list of variable importance methods to be set by the user. These can include any of the importance methods contained within the mlr package. The default is method = randomForest.}
}
\description{
Plots variable importance
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr ranger model:
library(mlr3)
aq_Task = TaskRegr$new(id = "airQ", backend = aq, target = "Ozone")
aq_lrn = lrn("regr.ranger", importance = "permutation")
aq_Mod <- lrn$train(aq_Task)

# Create matrix
myMat <- prepFunc(task = aq_Task, learner = aq_Lrn, model = aq_Mod)


# Create plot:
importancePlot(myMat)

}
