% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importancePlot.R
\name{importancePlot}
\alias{importancePlot}
\title{importancePlot}
\usage{
importancePlot(
  mat,
  plotType = "lollipop",
  minImp = NULL,
  maxImp = NULL,
  top = NULL,
  label,
  ...
)
}
\arguments{
\item{mat}{A matrix of values to be plotted. Either added by the user or created using the prepFunc() function.}

\item{minImp}{Minimum importance value to be displayed on the legend.}

\item{maxImp}{Maximum importance value to be displayed on the legend.}

\item{top}{A value set by the user to only display the top x amount variables.}

\item{label}{Only compatible with plotType = "barplot". If TRUE then the importance value is displayed at the end of each bar.}

\item{...}{Not currently implemented}

\item{type}{The type of plot to display, either "lollipop" (default), or "barplot".}
}
\description{
This function is used to plot just the variable importance.
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr ranger model:
library(mlr3)
library(mlr3learners)
library(ranger)
aq_Task = TaskRegr$new(id = "airQ", backend = aq, target = "Ozone")
aq_lrn = lrn("regr.ranger", importance = "permutation")
aq_Mod <- aq_lrn$train(aq_Task)

# Create matrix
myMat <- vividMatrix(task = aq_Task,  model = aq_Mod)

# Create plot:
plot(myMat, type = "importance")
}
