% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotImportance.R
\name{importancePlot}
\alias{importancePlot}
\title{importancePlot}
\usage{
importancePlot(mat, type = "lollipop", minImp = 0, maxImp = NULL, ...)
}
\arguments{
\item{mat}{A matrix of values to be plotted. Either added by the user or created using the prepFunc() function.}

\item{type}{The type of plot to display, either "lollipop" (default), "barplot", or "circleBar".}

\item{minImp}{Minimum importance value to be displayed on the legend.}

\item{maxImp}{Maximum importance value to be displayed on the legend.}

\item{...}{Not currently implemented}
}
\description{
Plots variable importance
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr ranger model:
library(mlr3)
library(mlr3learners)
library(ranger)
aq_Task = TaskRegr$new(id = "airQ", backend = aq, target = "Ozone")
aq_lrn = lrn("regr.ranger", importance = "permutation")
aq_Mod <- aq_lrn$train(aq_Task)

# Create matrix
myMat <- prepFunc(task = aq_Task, learner = aq_lrn, model = aq_Mod)


# Create plot:
importancePlot(myMat)

}
