% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotInteraction.R
\name{interactionPlot}
\alias{interactionPlot}
\title{interactionPlot}
\usage{
interactionPlot(task, model, type = "lollipop")
}
\arguments{
\item{task}{Task created from the mlr package, either regression or classification.}

\item{model}{Any machine learning model.}

\item{type}{The type of plot to display, either "lollipop" (default), "barplot", or "circleBar"}

\item{...}{Not currently implemented}
}
\description{
Plots a selected variable against all other variables in a model
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr random forest model:
library(mlr)
library(randomForest)
aqRgrTask  <- makeRegrTask(data = aq, target = "Ozone")
aqRegrLrn <- makeLearner("regr.randomForest")
aqMod <- train(aqRegrLrn, aqRgrTask)

# Create plot:
interactionPlot(aqRgrTask, aqMod)

}
