% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionPlot.R
\name{interactionPlot}
\alias{interactionPlot}
\title{interactionPlot}
\usage{
interactionPlot(
  model,
  data,
  interactionType = "ice",
  pal = rev(sequential_hcl(palette = "Blues 3", n = 11)),
  fitlims = NULL,
  gridSize = 10,
  normalize = FALSE,
  n_max = 1000,
  seed = NULL,
  sqrt = FALSE,
  type = "lollipop",
  title = NULL,
  label = FALSE
)
}
\arguments{
\item{model}{A machine learning model created from mlr3 task and learner.}

\item{interactionType}{Are measures based on Friedman's H statistic ("H") or on "ice" curves?}

\item{pal}{A vector of colors to show values, for use with scale_fill_gradientn}

\item{fitlims}{Specifies the fit range for the color map for interaction strength.}

\item{gridSize}{The size of the grid for evaluating the predictions.}

\item{normalize}{Should the variances explained be normalized? Default is FALSE.}

\item{n_max}{Maximum number of data rows to consider.}

\item{seed}{An integer random seed used for subsampling.}

\item{sqrt}{In order to reproduce Friedman's H statistic, resulting values are root transformed. Set to FALSE if squared values should be returned.}

\item{type}{The type of plot to display, either "lollipop" (default) or "barplot".}

\item{title}{A title to be placed on the plot}

\item{label}{Only compatible with plotType = "barplot". If TRUE then the interaction value is displayed at the end of each bar.}
}
\description{
Plots a selected variable against all other variables in a model
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr ranger model:
library(mlr3)
library(mlr3learners)
library(ranger)
aq_Task <- TaskRegr$new(id = "airQ", backend = aq, target = "Ozone")
aq_lrn <- lrn("regr.ranger", importance = "permutation")
aq_Mod <- aq_lrn$train(aq_Task)

# Create plot:
interactionPlot(aq_Task, aq_Mod)

}
