% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewHeatmap.R
\name{intHeatmap}
\alias{intHeatmap}
\title{Title intHeatmap}
\usage{
intHeatmap(
  task,
  model,
  method = "randomForest",
  plotly = FALSE,
  intLow = "floralwhite",
  intHigh = "dodgerblue4",
  impLow = "white",
  impHigh = "firebrick1",
  ...
)
}
\arguments{
\item{task}{Task created from the mlr package, either regression or classification.}

\item{model}{Any machine learning model.}

\item{method}{A list of variable importance methods to be set by the user. These can include any of the importance methods contained within the mlr package. The default is method = randomForest.}

\item{plotly}{If plotly = TRUE then it displays an interactive plotly plot.}

\item{intLow}{Colour, set by the user, to display low interaction strengths.}

\item{intHigh}{Colour, set by the user, to display high interaction strengths.}

\item{impLow}{Colour, set by the user, to display low importance values.}

\item{impHigh}{Colour, set by the user, to display high importance values.}
}
\value{
A heatmap style plot dispaying interaction strength on the off-diagonal and variable importance on the diagonal.
}
\description{
Title intHeatmap
}
