% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdpPairs.R
\name{pdpPairs}
\alias{pdpPairs}
\title{Title pdpPairs
name pdpPairs}
\usage{
pdpPairs(task, model, cols = rev(sequential_hcl(20, "Blues3")), ...)
}
\arguments{
\item{task}{Task created from the mlr package, either regression or classification.}

\item{model}{Any machine learning model.}

\item{cols}{Colour of partial dependece plot.}
}
\value{
A pairs style plot displaying the partial dependence.
}
\description{
Title pdpPairs
name pdpPairs
}
