% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeat.R
\name{plotHeat}
\alias{plotHeat}
\title{plotHeat}
\usage{
plotHeat(
  dinteraction,
  plotly = FALSE,
  top = NULL,
  title = "",
  intPal = rev(sequential_hcl(palette = "Blues 3", n = 11)),
  impPal = rev(sequential_hcl(palette = "Reds 3", n = 11)),
  minImp = NULL,
  maxImp = NULL,
  minInt = NULL,
  maxInt = NULL,
  angle = NULL,
  ...
)
}
\arguments{
\item{plotly}{If TRUE then an interactive plot is displayed.}

\item{top}{Returns the first part of the interaction matrix and resulting plot. Similar to head() function.}

\item{title}{Adds title to the plot.}

\item{intPal}{A colorspace colour palette to display the interaction values.}

\item{impPal}{A colorspace colour palette to display the importance values.}

\item{minImp}{Minimum importance value to be displayed on the legend.}

\item{maxImp}{Maximum importance value to be displayed on the legend.}

\item{minInt}{Minimum interaction strength to be displayed on the legend.}

\item{maxInt}{Maximum interaction strength to be displayed on the legend.}

\item{angle}{The angle to display the x-axis labels.}

\item{...}{Not currently implemented.}

\item{mat}{A matrix of values to be plotted. Either added by the user or created using the prepFunc() function.}
}
\description{
Plots a Heatmap-tyle display showingVariable Importance and Variable Interaction
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr ranger model:
library(mlr3)
library(mlr3learners)
library(ranger)
aq_Task <- TaskRegr$new(id = "airQ", backend = aq, target = "Ozone")
aq_lrn <- lrn("regr.ranger", importance = "permutation")
aq_Mod <- aq_lrn$train(aq_Task)

# Create matrix
myMat <- vividMatrix(task = aq_Task,  model = aq_Mod)

# Create plot:
plot(myMat, type = "heatMap")
}
