% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatMap.R
\name{plotHeatMap}
\alias{plotHeatMap}
\title{plotHeatMap}
\usage{
plotHeatMap(
  mat,
  plotly = FALSE,
  intLow = "floralwhite",
  intHigh = "dodgerblue4",
  impLow = "white",
  impHigh = "firebrick1",
  top = NULL,
  reorder = TRUE,
  minImp = 0,
  maxImp = NULL,
  minInt = 0,
  maxInt = NULL,
  ...
)
}
\arguments{
\item{mat}{A matrix of values to be plotted. Either added by the user or created using the prepFunc() function.}

\item{plotly}{If TRUE then an interactive plot is displayed.}

\item{intLow}{Colour, set by the user, to display low interaction strengths.}

\item{intHigh}{Colour, set by the user, to display high interaction strengths.}

\item{impLow}{Colour, set by the user, to display low importance values.}

\item{impHigh}{Colour, set by the user, to display high importance values.}

\item{top}{Returns the first part of the interaction matrix and resulting plot. Similar to head() function.}

\item{reorder}{If TRUE (default) uses DendSer to reorder the matrix of interactions and variable importances}

\item{minImp}{Minimum importance value to be displayed on the legend.}

\item{maxImp}{Maximum importance value to be displayed on the legend.
@param ... Not currently implemented}

\item{minInt}{Minimum interaction strength to be displayed on the legend.}

\item{maxInt}{Maximum interaction strength to be displayed on the legend.}
}
\description{
Plots a Heatmap-tyle display showingVariable Importance and Variable Interaction
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr ranger model:
library(mlr3)
library(mlr3learners)
library(ranger)
aq_Task = TaskRegr$new(id = "airQ", backend = aq, target = "Ozone")
aq_lrn = lrn("regr.ranger", importance = "permutation")
aq_Mod <- lrn$train(aq_Task)

#' # Create matrix
myMat <- prepFunc(task = aq_Task, learner = aq_Lrn, model = aq_Mod)

# Create plot:
plotHeatMap(myMat)

}
