% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepFunction.R
\name{prepFunc}
\alias{prepFunc}
\title{prepFunction}
\usage{
prepFunc(task, model, remove = FALSE, percentRemove = 0.5, parallel)
}
\arguments{
\item{task}{Task created from the mlr package, either regression or classification.}

\item{model}{Any machine learning model.}

\item{remove}{If TRUE then remove the variables with low interaction strength.}

\item{percentRemove}{The percentage of variables with low interaction strength to remove from the interaction calculation.}

\item{parallel}{If TRUE then the method is executed in parallel.}

\item{...}{Not currently implemented.}
}
\value{
A matrix of values
}
\description{
Creates a matrix displaying Variable importance on the diagonal
 and Variable Interaction on the off-diagonal.
}
\examples{
# Load in the data:
aq <- data.frame(airquality)
aq <- na.omit(aq)

# Run an mlr random forest model:
library(mlr)
library(randomForest)
aqRgrTask  <- makeRegrTask(data = aq, target = "Ozone")
aqRegrLrn <- makeLearner("regr.randomForest")
aqMod <- train(aqRegrLrn, aqRgrTask)

# Create graph:
plotNetwork(task = aqRgrTask, model = aqMod,
thresholdValue = 0, cluster = F)

}
